+++ src/utils/schemaGetters.js
@@ -151,3 +151,17 @@ export const getFieldConditions = (schema, modelName, fieldName) => {
 export const getFieldDisableCondition = (schema, modelName, fieldName) => {
   return R.propOr(null, 'disabled', getField(schema, modelName, fieldName))
 }
+
+export const getDropDownDisableCondition = (schema, modelName, fieldName) => {
+  return R.propOr(null, 'disabledDropDown', getField(schema, modelName, fieldName))
+}
+
+export const getOptionsOverride = ({schema, modelName, fieldName, options, formStack, value, modelStore }) => {
+  const disabledDropDownCond = getDropDownDisableCondition(schema, modelName, fieldName)
+  if (disabledDropDownCond) {
+    options = disabledDropDownCond({
+      schema, modelName, fieldName, options, formStack, value, modelStore
+    })
+  }
+  return options
+}